Check that an ami is findable via amazon's search
resource before we declare that the bake is complete. This avoid the
rather dangerous case where someone could do `klink deploy <service>
-m`, not specifiying the latest ami, assuming that the one they just
baked is available and resulting in them thinking they have deployed a
version that they haven't. Not sure why this is now an issue when it
wasn't before as ditto does check this. Must be a lag between the ami
being available and it being findable through a search. AMAZON HOUSE!